fix: rm try except into test
